KPV-1697 :: Generic AbstractCampaignCreatorService
